+++ crates/influxive-writer/tests/common/mod.rs
@@ -1,0 +1,3 @@
+pub mod telegraf_binaries;
+pub mod telegraf_influx_file_conf;
+pub mod telegraf_svc;